Nudge fuzzy finder icons
On Win + Linux, with a font-size > 13px, when resizing the window, it can happen that the icons are cut off on the left. Most likely due to some rounding error when converting em to px.

This commit adds an extra px to give a bit more buffer.

Fixes https://github.com/atom/atom/pull/8821